refactor: Refactor documentation

+++ src/commands/pr/list.rs
@@ -10,7 +10,7 @@ pub struct PullRequestListCommand;
 #[allow(dead_code)]
 impl PullRequestListCommand {
     /// 列出 PR
-    pub fn list(state: Option<String>, limit: Option<u32>) -> Result<()> {
+    pub fn list(state: Option<String>, limit: Option<usize>) -> Result<()> {
         log_break!('=', 40, "PR List");
         let provider = create_provider_auto()?;
